create trial balance table.

+++ lib/features/trialBalance/data/models/trial_balance_model.dart
@@ -5,7 +5,7 @@ class TrialBalanceModel {
   int? level;
   bool? beforCorD;
   double? beformony;
-  int? creditmony;
+  double? creditmony;
   double? depitmony;
   bool? creditORDepit;
   double? mony;
